Refactor and move UI table to widgets.py file
