+++ src/codeflare_sdk/__init__.py
@@ -14,7 +14,7 @@ from .cluster import (
     get_cluster,
     list_all_queued,
     list_all_clusters,
-    list_cluster_details,
+    view_clusters,
 )
 
 from .job import RayJobClient
