+++ src/codeflare_sdk/cluster/cluster.py
@@ -35,10 +35,6 @@ from ..utils.generate_yaml import (
 from ..utils.kube_api_helpers import _kube_api_error_handling
 from ..utils.generate_yaml import is_openshift_cluster
 
-import ipywidgets as widgets
-from IPython.display import display, HTML, Javascript
-import pandas as pd
-
 from .config import ClusterConfiguration
 from .model import (
     AppWrapper,
@@ -596,200 +592,6 @@ def get_current_namespace():  # pragma: no cover
         return active_context["context"]["namespace"]
     except KeyError:
         return None
-
-# format_status takes a RayCluster status and applies colors and icons based on the status.
-def format_status(status):
-    if status == RayClusterStatus.READY:
-        return '<span style="color: green;">Ready ✓</span>'
-    elif status == RayClusterStatus.SUSPENDED:
-        return '<span style="color: #007BFF;">Suspended ❄️</span>'
-    elif status == RayClusterStatus.FAILED:
-        return '<span style="color: red;">Failed ✗</span>'
-    elif status == RayClusterStatus.UNHEALTHY:
-        return '<span style="color: purple;">Unhealthy</span>'
-    elif status == RayClusterStatus.UNKNOWN:
-        return '<span style="color: purple;">Unknown</span>'
-    else:
-        return status
-
-def _fetch_cluster_data(namespace):
-    rayclusters = list_all_clusters(namespace, False)
-    names = [item.name for item in rayclusters]
-    namespaces = [item.namespace for item in rayclusters]
-    head_extended_resources = [
-        f"{list(item.head_extended_resources.keys())[0]}: {list(item.head_extended_resources.values())[0]}"
-        if item.head_extended_resources else "0"
-        for item in rayclusters
-    ]
-    worker_extended_resources = [
-        f"{list(item.worker_extended_resources.keys())[0]}: {list(item.worker_extended_resources.values())[0]}"
-        if item.worker_extended_resources else "0"
-        for item in rayclusters
-    ]
-    head_cpu_requests = [item.head_cpu_requests if item.head_cpu_requests else 0 for item in rayclusters]
-    head_cpu_limits = [item.head_cpu_limits if item.head_cpu_limits else 0 for item in rayclusters]
-    head_cpu_rl = [f"{requests}~{limits}" for requests, limits in zip(head_cpu_requests, head_cpu_limits)]
-    head_mem_requests = [item.head_mem_requests if item.head_mem_requests else 0 for item in rayclusters]
-    head_mem_limits = [item.head_mem_limits if item.head_mem_limits else 0 for item in rayclusters]
-    head_mem_rl = [f"{requests}~{limits}" for requests, limits in zip(head_mem_requests, head_mem_limits)]
-    worker_cpu_requests = [item.worker_cpu_requests if item.worker_cpu_requests else 0 for item in rayclusters]
-    worker_cpu_limits = [item.worker_cpu_limits if item.worker_cpu_limits else 0 for item in rayclusters]
-    worker_cpu_rl = [f"{requests}~{limits}" for requests, limits in zip(worker_cpu_requests, worker_cpu_limits)]
-    worker_mem_requests = [item.worker_mem_requests if item.worker_mem_requests else 0 for item in rayclusters]
-    worker_mem_limits = [item.worker_mem_limits if item.worker_mem_limits else 0 for item in rayclusters]
-    worker_mem_rl = [f"{requests}~{limits}" for requests, limits in zip(worker_mem_requests, worker_mem_limits)]
-    status = [item.status.name for item in rayclusters]
-
-    status = [format_status(item.status) for item in rayclusters]
-
-    data = {
-        "Name": names,
-        "Namespace": namespaces,
-        "Head GPUs": head_extended_resources,
-        "Worker GPUs": worker_extended_resources,
-        "Head CPU Req~Lim": head_cpu_rl,
-        "Head Memory Req~Lim": head_mem_rl,
-        "Worker CPU Req~Lim": worker_cpu_rl,
-        "Worker Memory Req~Lim": worker_mem_rl,
-        "status": status
-    }
-    return pd.DataFrame(data)
-
-
-def list_cluster_details(namespace: str): # TODO: or current namespace as default
-    global df
-    df = _fetch_cluster_data(namespace)
-
-    outputs = widgets.Output()
-    data_output = widgets.Output()
-    if df["Name"].empty:
-        print(f"No clusters found in the {namespace} namespace.")
-    else:
-        classification_widget = widgets.ToggleButtons(
-            options=df["Name"].tolist(), value=None,
-            description='Select an existing cluster:',
-        )
-
-        def on_cluster_click(change):
-            new_value = change["new"]
-            data_output.clear_output()
-            df = _fetch_cluster_data(namespace)
-            classification_widget.options = df["Name"].tolist()
-            with data_output:
-                display(HTML(df[df["Name"]==new_value][["Name", "Namespace", "Head GPUs", "Head CPU Req~Lim", "Head Memory Req~Lim", "Worker GPUs", "Worker CPU Req~Lim", "Worker Memory Req~Lim", "status"]].to_html(escape=False, index=False, border=2)))
-
-        classification_widget.observe(on_cluster_click, names="value")
-        display(widgets.VBox([classification_widget, data_output]))
-
-        def on_delete_button_clicked(b):
-            global df
-            cluster_name = classification_widget.value
-            namespace = df[df["Name"]==classification_widget.value]["Namespace"].values[0]
-            delete_cluster(cluster_name, namespace)
-
-            sleep(3) # wait for the cluster to be deleted
-            outputs.clear_output()
-            with outputs:
-                print(f"Cluster {cluster_name} in the {namespace} namespace was deleted successfully.")
-            # Refresh the dataframe
-            df = _fetch_cluster_data(namespace)
-            if df["Name"].empty:
-                classification_widget.close()
-                delete_button.close()
-                list_jobs_button.close()
-                ray_dashboard_button.close()
-                data_output.clear_output()
-                with data_output:
-                    print(f"No clusters found in the {namespace} namespace.")
-            else:
-                classification_widget.options = df["Name"].tolist()
-
-
-
-        # out Output widget is used to execute JavaScript code to open the Ray dashboard URL in a new browser tab
-        out = widgets.Output()
-        def on_ray_dashboard_button_clicked(b):
-            cluster_name = classification_widget.value
-            namespace = df[df["Name"]==classification_widget.value]["Namespace"].values[0]
-
-            cluster = Cluster(ClusterConfiguration(cluster_name, namespace))
-            dashboard_url = cluster.cluster_dashboard_uri()
-
-            outputs.clear_output()
-            with outputs:
-                print(f"Opening Ray Dashboard for {cluster_name} cluster:\n{dashboard_url}")
-            with out:
-                display(Javascript(f'window.open("{dashboard_url}", "_blank");'))
-
-        def on_list_jobs_button_clicked(b):
-            cluster_name = classification_widget.value
-            namespace = df[df["Name"]==classification_widget.value]["Namespace"].values[0]
-
-            cluster = Cluster(ClusterConfiguration(cluster_name, namespace))
-            dashboard_url = cluster.cluster_dashboard_uri()
-
-            outputs.clear_output()
-            with outputs:
-                print(f"Opening Ray Jobs Dashboard for {cluster_name} cluster:\n{dashboard_url}/#/jobs")
-            with out:
-                display(Javascript(f'window.open("{dashboard_url}/#/jobs", "_blank");'))
-
-        list_jobs_button = widgets.Button(
-                    description='View Jobs',
-                    icon='suitcase',
-                    tooltip="Open the Ray Job Dashboard"
-                )
-        list_jobs_button.on_click(on_list_jobs_button_clicked)
-
-        delete_button = widgets.Button(
-                    description='Delete Cluster',
-                    icon='trash',
-                    tooltip="Delete the selected cluster"
-                )
-        delete_button.on_click(on_delete_button_clicked)
-
-        ray_dashboard_button = widgets.Button(
-                    description='Open Ray Dashboard',
-                    icon='dashboard',
-                    tooltip="Open the Ray Dashboard in a new tab",
-                    layout=widgets.Layout(width='auto'),
-                )
-        ray_dashboard_button.on_click(on_ray_dashboard_button_clicked)
-
-        display(widgets.HBox([delete_button, list_jobs_button, ray_dashboard_button]), out, outputs)
-
-
-
-def delete_cluster(
-    cluster_name: str,
-    namespace: str, # TODO: get current namespace if not provided
-):
-    if _check_aw_exists(cluster_name, namespace):
-        try:
-            config_check()
-            api_instance = client.CustomObjectsApi(api_config_handler())
-            api_instance.delete_namespaced_custom_object(
-                group="workload.codeflare.dev",
-                version="v1beta2",
-                namespace=namespace,
-                plural="appwrappers",
-                name=cluster_name,
-            )
-        except Exception as e:
-            return _kube_api_error_handling(e)
-    else:
-        try:
-            config_check()
-            api_instance = client.CustomObjectsApi(api_config_handler())
-            api_instance.delete_namespaced_custom_object(
-                group="ray.io",
-                version="v1",
-                namespace=namespace,
-                plural="rayclusters",
-                name=cluster_name,
-            )
-        except Exception as e:
-            return _kube_api_error_handling(e)
 
 
 def get_cluster(
